Refactored id from int to Long Refactored class names

+++ src/main/java/com/myretail/api/model/Product.java
@@ -2,23 +2,23 @@ package com.myretail.api.model;
 
 public class Product {
 
-    private int id;
+    private Long id;
     private String name;
 
     public Product() {
     }
 
-    public Product(int id, String name, CurrentPrice currentPrice) {
+    public Product(Long id, String name, CurrentPrice currentPrice) {
         this.id = id;
         this.name = name;
         this.currentPrice = currentPrice;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
